Rename package to django_ratelimit
"Namespaces are one honking great idea -- let's do more of those!"

Follow the convention of most Django-focused libraries by namespacing
the main package. This fixes #214 and makes this library a better
neighbor.
