+++ django_ratelimit/tests.py
@@ -7,9 +7,10 @@ from django.test.utils import override_settings
 from django.utils.decorators import method_decorator
 from django.views.generic import View
 
-from ratelimit.decorators import ratelimit
-from ratelimit.exceptions import Ratelimited
-from ratelimit.core import get_usage, is_ratelimited, _split_rate, _get_ip
+from django_ratelimit.decorators import ratelimit
+from django_ratelimit.exceptions import Ratelimited
+from django_ratelimit.core import (get_usage, is_ratelimited,
+                                   _split_rate, _get_ip)
 
 
 rf = RequestFactory()
@@ -225,7 +226,8 @@ class RatelimitTests(TestCase):
             req.user = MockUser(authenticated=auth)
             return req
 
-        @ratelimit(key='user_or_ip', rate='ratelimit.tests.callable_rate')
+        @ratelimit(key='user_or_ip',
+                   rate='django_ratelimit.tests.callable_rate')
         def view(request):
             return request.limited
 
@@ -254,7 +256,7 @@ class RatelimitTests(TestCase):
         assert view(_req(auth=True))
 
     def test_callable_key_path(self):
-        @ratelimit(key='ratelimit.tests.mykey', rate='1/m')
+        @ratelimit(key='django_ratelimit.tests.mykey', rate='1/m')
         def view(request):
             return request.limited
 
@@ -607,7 +609,7 @@ class IpMetaTests(TestCase):
 
         assert '5.6.7.8' == _get_ip(req)
 
-    @override_settings(RATELIMIT_IP_META_KEY='ratelimit.tests.my_ip')
+    @override_settings(RATELIMIT_IP_META_KEY='django_ratelimit.tests.my_ip')
     def test_path_to_ip_key_callable(self):
         req = rf.get('/')
         req.META['REMOTE_ADDR'] = '1.2.3.4'
